% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxph2cph.R
\name{coxph2cph}
\alias{coxph2cph}
\title{Convert results of cox regression from coxph() to cph()}
\usage{
coxph2cph(fit)
}
\arguments{
\item{fit}{cox regression results of coxph()}
}
\value{
results of coxph()
}
\description{
Convert results of cox regression from coxph() in 'survival' package to cph() in 'rms' package.
}
\examples{
library(survival)
fit <- coxph(Surv(mpg,vs)~am+gear+carb,data=mtcars)
coxph2cph(fit)
}
