% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxph2cph.R
\name{cph2coxph}
\alias{cph2coxph}
\title{Convert results of cox regression from cph() to coxph()}
\usage{
cph2coxph(fit)
}
\arguments{
\item{fit}{cox regression results of cph()}
}
\value{
results of coxph()
}
\description{
Convert results of cox regression from cph() in 'rms' package to coxph() in 'stats' package.
}
\examples{
library(rms)
fit <- cph(formula = Surv(mpg, vs) ~ am + gear + carb, data = mtcars)
cph2coxph(fit)
}
