% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm.ols.R
\name{lm2ols}
\alias{lm2ols}
\title{Convert results of lm() to ols()}
\usage{
lm2ols(fit)
}
\arguments{
\item{fit}{reults of lm()}
}
\value{
results of ols()
}
\description{
Convert results of lm() from 'stats' package to ols() from 'rms' package.
}
\examples{
fit <- lm(mpg ~ disp,data=mtcars)
lm2ols(fit)
}
