% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit.lrm.R
\name{logit2lrm}
\alias{logit2lrm}
\title{Convert results of logistic regression from glm() to lrm()}
\usage{
logit2lrm(fit)
}
\arguments{
\item{fit}{logistic regression reults of glm()}
}
\value{
results of lrm()
}
\description{
Convert results of logistic regression from glm() in 'stats' package to lrm() in 'rms' package.
}
\examples{
fit <- glm(vs~mpg+cyl,data=mtcars, family = binomial(link = 'logit'))
logit2lrm(fit)
}
