% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit.lrm.R
\name{lrm2logit}
\alias{lrm2logit}
\title{Convert results of logistic regression from lrm() to glm()}
\usage{
lrm2logit(fit)
}
\arguments{
\item{fit}{logistic regression reults of lrm()}
}
\value{
results of glm()
}
\description{
Convert results of logistic regression from lrm() in 'rms' package to glm() in 'stats' package.
}
\examples{
library(rms)
fit <- lrm(vs ~ mpg + cyl, data = mtcars)
lrm2logit(fit)
}
