% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm.ols.R
\name{ols2lm}
\alias{ols2lm}
\title{Convert results of ols() to lm()}
\usage{
ols2lm(fit)
}
\arguments{
\item{fit}{reults of ols()}
}
\value{
results of lm()
}
\description{
Convert results of ols() from 'rms' package to lm() from 'stats' package.
}
\examples{
library(rms)
fit <- ols(mpg ~ disp^2,data=mtcars)
ols2lm(fit)
}
